Organizing files in separate folders.
