+++ compile_linux_version_with_gcc.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-g++ -pthread -std=c++17 black_marlin.cpp black_marlin.hpp httplib.h main.cpp http_request_handler.cpp http_request_handler.hpp server_configs.hpp server_configs.cpp posix_pathfinder.hpp posix_pathfinder.cpp util.hpp util.cpp server_connection_info.hpp -o black_marlin.file
+g++ -pthread -std=c++17 BlackMarlin/black_marlin.cpp BlackMarlin/black_marlin.hpp HTTP/httplib.h main.cpp HTTP/http_request_handler.cpp HTTP/http_request_handler.hpp ServerConfigs/server_configs.hpp ServerConfigs/server_configs.cpp ServerConfigs/posix_pathfinder.hpp ServerConfigs/posix_pathfinder.cpp Util/util.hpp Util/util.cpp HTTP/server_connection_info.hpp -o black_marlin.file
